requirements.txt
requests
feedparser
newsapi-python
nltk
torch
transformers
streamlit
schedule
pandas
